// shopify-embed-app/libs/webhooks/order/order.types.ts
export interface Staff {
  _id: string;
  fullname: string;
  avatar: string;
  position: string;
  anyAvailable?: boolean;
}

export interface Data {
  start: string;
  end: string;
  staff: Staff;
  timeZone: string;
}
// ...
export interface ShopMoney11 {
  amount?: string;
  currency_code: string;
}

export interface PresentmentMoney11 {
  amount: string;
  currency_code: string;
}

export interface PriceSet {
  shop_money: ShopMoney11;
  presentment_money: PresentmentMoney11;
}

export interface Property {
  name: string;
  value: string;
}

export interface ShopMoney12 {
  amount: string;
  currency_code: string;
}

export interface PresentmentMoney12 {
  amount: string;
  currency_code: string;
}

export interface TotalDiscountSet {
  shop_money: ShopMoney12;
  presentment_money: PresentmentMoney12;
}

export interface ShopMoney13 {
  amount: string;
  currency_code: string;
}

export interface PresentmentMoney13 {
  amount: string;
  currency_code: string;
}

export interface AmountSet {
  shop_money: ShopMoney13;
  presentment_money: PresentmentMoney13;
}

export interface DiscountAllocation {
  amount: string;
  amount_set: AmountSet;
  discount_application_index: number;
}
// ...
export interface LineItem {
  id: number;
  admin_graphql_api_id: string;
  fulfillable_quantity: number;
  fulfillment_service: string;
  fulfillment_status?: any;
  gift_card: boolean;
  grams: number;
  name: string;
  price: string;
  price_set: PriceSet;
  product_exists: boolean;
  product_id: number;
  properties: Property[];
  quantity: number;
  requires_shipping: boolean;
  sku: string;
  taxable: boolean;
  title: string;
  total_discount: string;
  total_discount_set: TotalDiscountSet;
  variant_id: number;
  variant_inventory_management?: any;
  variant_title?: any;
  vendor: string;
  tax_lines: any[];
  duties: any[];
  discount_allocations: DiscountAllocation[];
}